refactor: new gui
