+++ src/main/java/me/kiriyaga/nami/feature/gui/newgui/base/DataPanelComponent.java
@@ -1,0 +1,4 @@
+package me.kiriyaga.nami.feature.gui.newgui.base;
+
+public class DataPanelComponent {
+}
