+++ src/main/java/me/kiriyaga/nami/feature/gui/newgui/widget/ButtonWidget.java
@@ -1,4 +1,4 @@
-package me.kiriyaga.nami.feature.gui.base;
+package me.kiriyaga.nami.feature.gui.newgui.base;
 
 import me.kiriyaga.nami.feature.module.impl.client.ClickGuiModule;
 import me.kiriyaga.nami.feature.module.impl.client.ColorModule;
@@ -9,7 +9,7 @@ import net.minecraft.text.Text;
 import java.awt.*;
 
 import static me.kiriyaga.nami.Nami.*;
-import static me.kiriyaga.nami.feature.gui.base.GuiConstants.toRGBA;
+import static me.kiriyaga.nami.feature.gui.oldgui.base.GuiConstants.toRGBA;
 
 public class ButtonWidget {
     private final String label;
