+++ src/main/java/me/kiriyaga/nami/feature/gui/oldgui/components/SettingPanel.java
