+++ src/main/java/me/kiriyaga/nami/feature/gui/oldgui/settings/EnumSettingRenderer.java
