+++ src/main/java/me/kiriyaga/nami/feature/gui/oldgui/settings/IntSettingRenderer.java
